Definately an interesting commit overall.
Here we replace an old and failing CRC cloaking system (crackable), with
one that uses MD5. The MD5 version is uncrackable. Pretty soon, hostkeys
will be reimplemented. This implementation is similar to the one found in
UnrealIRCd, but is definately not the same. :-P

Cloak keys will be reimplemented soon. :-P.